shift to shinydashboard for arranging interactive content
This update mostly affects the sidebar, and even there, most of the
content is similar. Most changes are internal, and relate to how input
data is used within the server function. Only major user-visible change
is the removal of separate ‘article’ and ‘topic’ subplots; now only one
figure is shown.