switch to restyle for updates to plot colors
Major change in this version is to ensure that plot is not redrawn
every time a point is selected/excluded or recolored. Minor changes
include option to manually alter plot height, additional export
options, and use of shinycssloaders to show progress bar when plots are
processing. This is an in-progress update; start_review_window still
not completely functional.