update to row selection in topic modelling
Topic models require >0 terms to run, but the line to select these rows
was too late in the code. Ergo when data were supplied that lacked
title/abstracts, the input and modelled results had differing numbers
of rows. Appears to only be a problem for the input and not subsequent
re-calculations of the topic model.