Restructure preload_screen_abstracts
Previously this function built the app then saved it as a .RData file. This led to problems where paths created by the app on one machine were not transferrable to a second machine. The new solution is to create an S3 object which provides parameters to revtools, which builds the app on the second machine.